[Refactor](package.json): change module field to .mjs
[Refactor](json): update imports to use pointer sub-module
