+++ packages/aileron/development-helper/package.json
@@ -11,8 +11,8 @@
     "email": "lunox273@gmail.com"
   },
   "sideEffects": false,
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.esm.js",
+  "main": "dist/index.cjs",
+  "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "files": [
     "dist",
