+++ packages/canard/schema-form/package.json
@@ -16,12 +16,12 @@
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "import": "./dist/index.esm.js",
-      "require": "./dist/index.cjs.js"
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
     }
   },
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.esm.js",
+  "main": "dist/index.cjs",
+  "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "files": [
     "dist",
@@ -105,11 +105,11 @@
   },
   "size-limit": [
     {
-      "path": "dist/index.cjs.js",
+      "path": "dist/index.cjs",
       "limit": "20 KB"
     },
     {
-      "path": "dist/index.esm.js",
+      "path": "dist/index.mjs",
       "limit": "20 KB"
     }
   ]
